# Improve screenshot performance