--- a/navigation-fragments/src/main/java/dev/hotwire/navigation/util/HotwireViewScreenshotHolder.kt
+++ b/navigation-fragments/src/main/java/dev/hotwire/navigation/util/HotwireViewScreenshotHolder.kt
@@ -13,7 +13,6 @@ import dev.hotwire.navigation.logging.logEvent
 import dev.hotwire.navigation.views.HotwireView
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.suspendCancellableCoroutine
-import java.security.AccessController.getContext
 
 internal class HotwireViewScreenshotHolder {
     private var bitmap: Bitmap? = null
@@ -47,7 +46,9 @@ internal class HotwireViewScreenshotHolder {
             val start = System.currentTimeMillis()
             val window = hotwireView.getActivity()?.window
 
-            if (window == null || !hotwireView.isLaidOut || !hasEnoughMemoryForScreenshot() || (hotwireView.width <= 0 || hotwireView.height <= 0)) {
+            if (window == null || !hotwireView.isLaidOut || !hasEnoughMemoryForScreenshot() ||
+                hotwireView.width <= 0 || hotwireView.height <= 0
+            ) {
                 if (continuation.isActive) {
                     continuation.resume(null, null)
                 }
@@ -61,9 +62,7 @@ internal class HotwireViewScreenshotHolder {
 
             try {
                 PixelCopy.request(
-                    window,
-                    rect,
-                    bitmap,
+                    window, rect, bitmap,
                     { result ->
                         if (result == PixelCopy.SUCCESS) {
                             logEvent(