This PR replaces the original `View.drawToBitmap` way of capturing screenshots for HotwireView with a modern and more performant [PixelCopy API](https://developer.android.com/reference/android/view/PixelCopy).

Along with this, the screenshot properties have been extracted to a `HotwireScreenshotHolder` to improve code organisation.

----

Using the old canvas -> bitmap drawing approach produced wildly inconsistent times to create a bitmap. Sometimes taking over 300-450ms:
```
viewScreenshotCreated ............. [size: 1280x2172, duration: 436ms]
viewScreenshotCreated ............. [size: 1280x2172, duration: 48ms]
viewScreenshotCreated ............. [size: 1280x2172, duration: 300ms]
viewScreenshotCreated ............. [size: 1280x2172, duration: 39ms]
viewScreenshotCreated ............. [size: 1280x2172, duration: 155ms]
viewScreenshotCreated ............. [size: 1280x2172, duration: 43ms]
viewScreenshotCreated ............. [size: 1280x2172, duration: 32ms]
```

Using the `PixelCopy` API is much faster and more consistent, typically under 20ms to produce a bitmap:
```
viewScreenshotCreated ............. [size: 1280x2172, duration: 21ms]
viewScreenshotCreated ............. [size: 1280x2172, duration: 12ms]
viewScreenshotCreated ............. [size: 1280x2172, duration: 13ms]
viewScreenshotCreated ............. [size: 1280x2172, duration: 11ms]
viewScreenshotCreated ............. [size: 1280x2172, duration: 12ms]
viewScreenshotCreated ............. [size: 1280x2172, duration: 12ms]
viewScreenshotCreated ............. [size: 1280x2172, duration: 12ms]
viewScreenshotCreated ............. [size: 1280x2172, duration: 9ms]
```